% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRelease.R
\name{buildRelease}
\alias{buildRelease}
\title{Create OMOP CDM SQL files}
\usage{
buildRelease(
  cdmVersion = listSupportedVersions(),
  targetDialect = listSupportedDialects()
)
}
\arguments{
\item{cdmVersion}{The version of the CDM you are creating, e.g. 5.3, 5.4.
Defaults to all supported CDM versions.}

\item{targetDialect}{The target dialect}
}
\description{
Writes DDL, ForeignKey, PrimaryKey and index SQL files for given cdmVersion
and targetDialect to the 'ddl' folder in current working directory.
}
