% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonDataModel-package.R
\docType{package}
\name{CommonDataModel-package}
\alias{CommonDataModel}
\alias{CommonDataModel-package}
\title{CommonDataModel: OMOP CDM DDL and Documentation Generator}
\description{
Generates the scripts required to create an Observational Medical Outcomes Partnership (OMOP) Common Data Model (CDM) database and associated documentation for supported database platforms. Leverages the 'SqlRender' package to convert the Data Definition Language (DDL) script written in parameterized Structured Query Language (SQL) to the other supported dialects.
}
\author{
\strong{Maintainer}: Clair Blacketer \email{mblacke@its.jnj.com}

Authors:
\itemize{
  \item Clair Blacketer \email{mblacke@its.jnj.com}
}

}
\keyword{internal}
