% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createForeignKeys.R
\name{createForeignKeys}
\alias{createForeignKeys}
\title{Create a foreign key sql script from two csv files that detail the OMOP CDM Specifications.}
\usage{
createForeignKeys(cdmVersion = cdmVersion)
}
\arguments{
\item{cdmVersionNum}{The version of the CDM you are creating, e.g. 5.3.1}
}
\description{
Create a foreign key sql script from two csv files that detail the OMOP CDM Specifications.
}
