% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPkFromFile.R
\name{createPkFromFile}
\alias{createPkFromFile}
\title{Create a primary key sql script from two csv files that detail the OMOP CDM Specifications.}
\usage{
createPkFromFile(cdmVersionNum = cdmVersion,
  cdmFieldCsvLoc = "inst/csv/OMOP_CDMv5.3.1_Field_Level.csv",
  outputFile = paste0("inst/sql/sql_server/OMOP CDM pk ", cdmVersion,
  " ", Sys.Date(), ".sql"))
}
\arguments{
\item{cdmVersionNum}{The version of the CDM you are creating}

\item{cdmFieldCsvLoc}{The location of the csv file with the CDM field information. This is defaulted to "inst/csv/OMOP_CDMv5.3.1_Field_Level.csv".
If a new version of this file was committed to the CDM repository the package automatically will grab it and place it in "inst/csv/".}

\item{outputFile}{The name of the output primary key sql file. This is defaulted to a location in the inst/sql/sql server folder and named with today's date and CDM Version.}
}
\description{
Create a primary key sql script from two csv files that detail the OMOP CDM Specifications.
}
