% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrimaryKeys.R
\name{createPrimaryKeys}
\alias{createPrimaryKeys}
\title{Create a primary key sql script from two csv files that detail the OMOP CDM Specifications.}
\usage{
createPrimaryKeys(cdmVersion = cdmVersion)
}
\arguments{
\item{cdmVersionNum}{The version of the CDM you are creating, e.g. 5.3}
}
\description{
Create a primary key sql script from two csv files that detail the OMOP CDM Specifications.
}
