% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSupportedVersions.R
\name{listSupportedDialects}
\alias{listSupportedDialects}
\title{List RDBMS dialects supported by this package}
\usage{
listSupportedDialects()
}
\value{
A list containing the supported Structured Query Language (SQL) dialects.
}
\description{
List RDBMS dialects supported by this package
}
