% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiParser.R
\name{parseWiki}
\alias{parseWiki}
\title{Parse Wiki files}
\usage{
parseWiki(mdFilesLocation, output_file)
}
\arguments{
\item{mdFilesLocation}{Path to the root folder of the Wiki repository.}

\item{output_file}{Path to where the output CSV file should be written.}
}
\description{
Parses all .md files in the specified location (or any subfolders), extracting definitions
of the Common Data Model.
}
