% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDDL.R
\name{writeDDL}
\alias{writeDDL}
\title{Write DDL script}
\usage{
writeDDL(targetdialect, cdmVersion, cdmDatabaseSchema = "@cdmDatabaseSchema")
}
\arguments{
\item{targetdialect}{The dialect of the target database. Choices are "oracle", "postgresql", "pdw", "redshift", "impala", "netezza", "bigquery", "sql server"}

\item{cdmVersion}{The version of the CDM for which you are creating the DDL. e.g. 5.3.1}

\item{cdmDatabaseSchema}{The schema of the CDM instance where the DDL will be run. For example, this would be "ohdsi.dbo" when testing on sql server.
Defaults to "@cdmDatabaseSchema"}
}
\description{
Write DDL script
}
