% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeIndex.R
\name{writeIndex}
\alias{writeIndex}
\title{Write Index script}
\usage{
writeIndex(targetdialect, cdmVersion, cdmDatabaseSchema, sqlFilename)
}
\arguments{
\item{targetdialect}{The dialect of the target database. Choices are "oracle", "postgresql", "pdw", "redshift", "impala", "netezza", "bigquery", "sql server"}

\item{cdmVersion}{The version of the CDM that you are creating the primary keys for}

\item{cdmDatabaseSchema}{The name of the schema where the cdm sits.}

\item{sqlFilename}{The name of the file that should be rendered and translated to a different dbms. This is either "OMOP CDM indices v5_3_1.sql" or "OMOP CDM indices v6_0.sql".}
}
\description{
Write Index script
}
