% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePrimaryKeys.R
\name{writePrimaryKeys}
\alias{writePrimaryKeys}
\title{Write primary key script}
\usage{
writePrimaryKeys(
  targetdialect,
  cdmVersion,
  cdmDatabaseSchema = "@cdmDatabaseSchema"
)
}
\arguments{
\item{targetdialect}{The dialect of the target database. Choices are "oracle", "postgresql", "pdw", "redshift", "impala", "netezza", "bigquery", "sql server"}

\item{cdmVersion}{The version of the CDM that you are creating the primary keys for}

\item{cdmDatabaseSchema}{The name of the schema where the cdm sits. Defaults to "@cdmDatabaseSchema"}
}
\description{
Write primary key script
}
