% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSupportedVersions.R
\name{listSupportedVersions}
\alias{listSupportedVersions}
\title{List CDM versions supported by this package}
\usage{
listSupportedVersions()
}
\value{
A character vector containing the support CDM versions in {major}.{minor} format.
}
\description{
List CDM versions supported by this package
}
