% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeIndex.R
\name{writeIndex}
\alias{writeIndex}
\title{Write Index script}
\usage{
writeIndex(targetdialect, cdmVersion, cdmDatabaseSchema = "@cdmDatabaseSchema")
}
\arguments{
\item{targetdialect}{The dialect of the target database. Choices are "oracle", "postgresql", "pdw", "redshift", "impala", "netezza", "bigquery", "sql server"}

\item{cdmVersion}{The version of the CDM that you are creating the indices for. e.g. 5.3.1}

\item{cdmDatabaseSchema}{The name of the schema where the cdm sits. Defaults to "@cdmDatabaseSchema"}
}
\description{
Write Index script
}
